Add abstract class SuggestionList
